KTY-65 address Ask Kilian QC findings
